Fix tests

+++ server/src/main/java/app/ister/server/EventHandler.java
@@ -1,5 +1,7 @@
 package app.ister.server;
 
+import app.ister.server.enums.EventType;
+import app.ister.server.eventHandlers.Handle;
 import app.ister.server.eventHandlers.HandleEpisodeFound;
 import app.ister.server.eventHandlers.HandleMediaFileFound;
 import app.ister.server.eventHandlers.HandleNewDirectoriesScanRequested;
@@ -12,28 +14,29 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-@Component
+import java.util.Arrays;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.function.Function;
+import java.util.stream.Collectors;
+
+@Service
 @Slf4j
 public class EventHandler {
-    @Autowired
-    private ServerEventRepository serverEventRepository;
-    @Autowired
-    private HandleNewDirectoriesScanRequested handleNewDirectoriesScanRequested;
-    @Autowired
-    private HandleShowFound handleShowFound;
-    @Autowired
-    private HandleEpisodeFound handleEpisodeFound;
-    @Autowired
-    private HandleMediaFileFound handleMediaFileFound;
-    @Autowired
-    private HandleNfoFileFound handleNfoFileFound;
-    @Autowired
-    private HandleSubtitleFileFound handleSubtitleFileFound;
-
+    private final ServerEventRepository serverEventRepository;
+    private final Map<EventType, Handle> handles;
 
     private boolean handling;
+
+    public EventHandler(final ServerEventRepository serverEventRepository,
+                        final Handle... handles) {
+        this.serverEventRepository = serverEventRepository;
+        this.handles = Arrays.stream(handles).collect(Collectors.toMap(Handle::handles, Function.identity()));
+
+    }
 
     @Scheduled(cron = "*/5 * * * * *")
     @Transactional
@@ -41,20 +44,18 @@ public class EventHandler {
         if (!handling) {
             handling = true;
             serverEventRepository.findAllByFailedIsFalse(Pageable.ofSize(300)).forEach(serverEventEntity -> {
-                log.debug("Handling event: {}, for type: {}", serverEventEntity.getPath(), serverEventEntity.getEventType());
-                Boolean successful = switch (serverEventEntity.getEventType()) {
-                    case NEW_DIRECTORIES_SCAN_REQUEST -> handleNewDirectoriesScanRequested.handle(serverEventEntity);
-                    case SHOW_FOUND -> handleShowFound.handle(serverEventEntity);
-                    case EPISODE_FOUND -> handleEpisodeFound.handle(serverEventEntity);
-                    case MEDIA_FILE_FOUND -> handleMediaFileFound.handle(serverEventEntity);
-                    case NFO_FILE_FOUND -> handleNfoFileFound.handle(serverEventEntity);
-                    case SUBTITLE_FILE_FOUND -> handleSubtitleFileFound.handle(serverEventEntity);
-                };
-                if (successful) {
-                    serverEventRepository.delete(serverEventEntity);
+                Handle handle = handles.get(serverEventEntity.getEventType());
+                if (handle != null) {
+                    log.debug("Handling event: {}, for type: {}", serverEventEntity.getPath(), serverEventEntity.getEventType());
+                    Boolean successful = handle.handle(serverEventEntity);
+                    if (successful) {
+                        serverEventRepository.delete(serverEventEntity);
+                    } else {
+                        serverEventEntity.setFailed(true);
+                        serverEventRepository.save(serverEventEntity);
+                    }
                 } else {
-                    serverEventEntity.setFailed(true);
-                    serverEventRepository.save(serverEventEntity);
+                    log.debug("No handler found for event: {}, for type: {}", serverEventEntity.getPath(), serverEventEntity.getEventType());
                 }
             });
             handling = false;
